polished routes and controllers

+++ controllers/userControl.js
@@ -13,7 +13,9 @@ module.exports = {
     // get single user
     async whichThinker(req, res) {
         try {
-            const user = await User.findOne({ _id: req.params. });
+            const user = await User.findOne(
+                { _id: req.params.userId }
+                ).populate('thoughts').populate('friends');
 
             if (!user) {
                 return res.status(404).json({ message: 'No Free Thinker Located' });
@@ -29,6 +31,7 @@ module.exports = {
             const user = await User.create(req.body);
             res.json(user);
         } catch (error) {
+            console.log(error);
             res.status(500).json(error);
         }
     },
@@ -36,7 +39,7 @@ module.exports = {
     async changedThinker(req, res) {
         try {
             const user = await User.findOneAndUpdate(
-                { _id: req.params. },
+                { _id: req.params.userId },
                 { $set: req.body },
                 { runValidators: true, new: true }
                 );
@@ -51,7 +54,7 @@ module.exports = {
     // delete user, bonus: remove associated thoughts
     async lostThinker(req, res) {
         try {
-            const user = await User.deleteOne({ _id: req.params. });
+            const user = await User.deleteOne({ _id: req.params.userId });
             if (!user) {
                 return res.status(404).json({ message: 'No Free Thinker Located' });
             }
@@ -61,5 +64,41 @@ module.exports = {
         } catch (error) {
             res.status(500).json(error);
         }
-    }
+    },
+    // add friend
+    async newFriend(req, res) {
+        try {
+            const user = await User.findOneAndUpdate(
+              { _id: req.params.userId },
+              { $addToSet: { friends: req.params.friendId } },
+              { runValidators: true, new: true }
+            );
+      
+            if (!user) {
+              return res.status(404).json({ message: 'No user with this id!' });
+            }
+      
+            res.json(user);
+          } catch (err) {
+            res.status(500).json(err);
+          }
+    },
+    // delete friend
+    async byeFriend(req, res) {
+        try {
+            const user = await User.findOneAndUpdate(
+              { _id: req.params.userId },
+              { $pull: { friends: req.params.friendId } },
+              { runValidators: true, new: true }
+            );
+      
+            if (!user) {
+              return res.status(404).json({ message: 'No friend with this id!' });
+            }
+      
+            res.json(user);
+          } catch (err) {
+            res.status(500).json(err);
+          }
+    },
 };
